Adds request_utils.py
