+++ usecase/create_dataset_entity_for_gtfs_metadata.py
@@ -1,5 +1,10 @@
-from utilities.validators import validate_gtfs_representation, validate_api_url
 import requests
+from utilities.request_utils import (
+    create_wikibase_item_claim_string,
+    create_regular_claim_string,
+    generate_api_csrf_token,
+)
+from utilities.validators import validate_gtfs_representation, validate_api_url
 
 
 def create_data(metadata):
@@ -24,49 +29,6 @@ def create_data(metadata):
         }}"""
 
 
-def create_wikibase_item_claim_string(property_id, entity_id, rank):
-    value = f"""{{
-                "entity-type":"item", 
-                "id":"{entity_id}"
-            }}"""
-    return create_claim_string(
-        property_id, value, rank, "wikibase-entityid", "wikibase-item"
-    )
-
-
-def create_regular_claim_string(property_id, value, rank):
-    value = f'"{value}"'
-    return create_claim_string(property_id, value, rank, "string", "string")
-
-
-def create_claim_string(property_id, value, rank, type, datatype):
-    return f"""
-        "{property_id}":[
-            {{
-                "mainsnak": {{
-                    "snaktype": "value",
-                    "property": "{property_id}",
-                    "datavalue": {{
-                        "value": {value},
-                        "type": "{type}"
-                    }},
-                    "datatype": "{datatype}"
-                }},
-                "type": "statement",
-                "rank": "{rank}"
-            }}
-        ]
-        """
-
-
-def create_geographical_claim_string(property_id, value, rank):
-    raise NotImplementedError
-
-
-def create_geographical_item(rank, value):
-    raise NotImplementedError
-
-
 def create_dataset_entity_for_gtfs_metadata(gtfs_representation, api_url):
     """Create a dataset entity for a new dataset version on the Database.
     :param gtfs_representation: The representation of the GTFS dataset to process.
@@ -77,44 +39,10 @@ def create_dataset_entity_for_gtfs_metadata(gtfs_representation, api_url):
     validate_gtfs_representation(gtfs_representation)
     metadata = gtfs_representation.metadata
 
-    # Step 1: GET request to fetch login token
-    PARAMS_LOGIN_TOKEN = {
-        "action": "query",
-        "meta": "tokens",
-        "type": "login",
-        "format": "json",
-    }
-
-    api_response = requests.get(api_url, params=PARAMS_LOGIN_TOKEN)
-    api_response.raise_for_status()
-    response_data = api_response.json()
-
-    LOGIN_TOKEN = response_data["query"]["tokens"]["logintoken"]
-
-    # Step 2: POST request to log in. Use of main account for login is not
-    # supported. Obtain credentials via Special:BotPasswords
-    # (https://www.mediawiki.org/wiki/Special:BotPasswords) for lgname & lgpassword
-    PARAMS_LOGIN = {
-        "action": "login",
-        "lgname": "MaximeArmstrong",
-        "lgpassword": "j9-sb4E7AKFTu8iVp93erLuYqqNeKv3ZyvN7cNKG",
-        "lgtoken": LOGIN_TOKEN,
-        "format": "json",
-    }
-
-    api_response = requests.post(api_url, data=PARAMS_LOGIN_TOKEN)
-    api_response.raise_for_status()
-
-    # Step 3: GET request to fetch CSRF token
-    PARAMS_CSRF_TOKEN = {"action": "query", "meta": "tokens", "format": "json"}
-
-    api_response = requests.get(api_url, params=PARAMS_CSRF_TOKEN)
-    api_response.raise_for_status()
-    response_data = api_response.json()
-
-    CSRF_TOKEN = response_data["query"]["tokens"]["csrftoken"]
+    CSRF_TOKEN = generate_api_csrf_token(api_url)
 
     # Step 4: POST request to edit a page
+    # print(create_data(metadata))
     PARAMS_ENTITY_CREATION = {
         "action": "wbeditentity",
         "new": "item",
